requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.0
antlr4-python3-runtime==4.8
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.5
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.11.0
diffusers==0.28.0
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
huggingface-hub==0.23.2
idna==3.4
imageio==2.31.4
importlib-metadata==7.0.1
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.5
kornia==0.6.12
lazy_loader==0.3
lightning-utilities==0.10.1
lit==17.0.1
lpips==0.1.4
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.1
ml_collections==0.1.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.25.3
pudb==2024.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==1.0.3
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.66.2
triton==2.0.0
tzdata==2023.3
urllib3==2.0.5
urwid==2.6.8
urwid_readline==0.14
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
